pluginhandler: add builtin functions to scriptlets
There is the desire for scriptlets to be able to communicate with
Snapcraft, e.g. to request that it use a specific name, or version, or
do a specific action, such as fetch stage packages, or run a lifecycle
step.

Add the basic framework for this communication. Add a new function
available to scriptlets called `snapcraft_build` which, if called, will
run the plugin's build step.

Resolve #1675
LP: #1751749

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>